Adjust Incentives to render CMS data
